requirements.txt
tensorflow_gpu==1.4.0
pandas==0.23.1
tqdm==4.23.4
numpy==1.14.5
